+++ CHANGELOG.md
@@ -1,18 +1,21 @@
 # Changelog
 
 ## 1.2.0
+- added item stack iota and patterns
+- added status effect patterns
 - added random achievements for fun
 - added z-axis rotation for specks
 - added energized modifier to mage blocks to make them emit Redstone power
 - added Displace spell for circles that lets them teleport entities for cheap
+- changed Conjure Speck to push the speck entity to the stack
 - fixed accidental swapping of Sloth and Racer's Purification
 - fixed being able to use Recharge Item to recharge lamps
 - fixed conjured staves not casting properly
 - fixed mage block breaking particles
-- fixed speck text rendering
 - fixed speck pattern saving
+- fixed speck text rendering
 - remove range requirement for altering specks
-- updated some documentation
+- updated documentation
 
 ## 1.1.0
 - added mishap to grimoire patterns if you are not holding a grimoire in your offhand